+++ certbot/certbot/tests/util.py
@@ -395,8 +395,9 @@ class ConfigTestCase(TempDirTestCase):
     def setUp(self) -> None:
         super().setUp()
         self.config = configuration.NamespaceConfig(
-            mock.MagicMock(**constants.CLI_DEFAULTS)
+            mock.MagicMock(**constants.CLI_DEFAULTS),
         )
+        self.config.set_argument_sources({})
         self.config.namespace.verb = "certonly"
         self.config.namespace.config_dir = os.path.join(self.tempdir, 'config')
         self.config.namespace.work_dir = os.path.join(self.tempdir, 'work')